Fix tests (#238)
* Begin reworking tests

Progress on #237

* Respect XDG Base Directory spec (#239)

Fix #236

* Fix config creation bug

* Get all tests to pass

Fix #237

* Apparently travis VMs don't come with /tmp created already

* Wrong diagnosis of the issue

+++ tests/test_copies.py
@@ -2,13 +2,11 @@ import os
 import sys
 import pytest
 import shutil
+from .test_utils import setup_env_vars, unset_env_vars, BACKUP_DEST_DIR, FAKE_HOME_DIR, DIRS
 sys.path.insert(0, "../shallow_backup")
 from shallow_backup.utils import copy_dir_if_valid
 
-DIR_TO_BACKUP = 'shallow-backup-test-copy-dir'
-BACKUP_DIR = 'shallow-backup-test-copy-backup-dir'
-TEST_TEXT_FILE = 'test-file.txt'
-DIRS = [DIR_TO_BACKUP, BACKUP_DIR]
+TEST_TEXT_FILE = os.path.join(FAKE_HOME_DIR, 'test-file.txt')
 
 
 class TestCopyMethods:
@@ -18,37 +16,39 @@ class TestCopyMethods:
 
     @staticmethod
     def setup_method():
-        for directory in DIRS:
-            try:
-                os.mkdir(directory)
-            except FileExistsError:
-                shutil.rmtree(directory)
-                os.mkdir(directory)
-        f = open(TEST_TEXT_FILE, "w+")
-        f.close()
+        setup_env_vars()
+        try:
+            os.mkdir(FAKE_HOME_DIR)
+        except FileExistsError:
+            shutil.rmtree(FAKE_HOME_DIR)
+            os.mkdir(FAKE_HOME_DIR)
+        print(f"Created {TEST_TEXT_FILE}")
+        open(TEST_TEXT_FILE, "w+").close()
 
     @staticmethod
     def teardown_method():
         for directory in DIRS:
-            shutil.rmtree(directory)
-        os.remove(TEST_TEXT_FILE)
+            if os.path.isdir(directory):
+                shutil.rmtree(directory)
+        unset_env_vars()
 
     def test_copy_dir(self):
         """
         Test that copying a directory works as expected
         """
         # TODO: Test that all subfiles and folders are copied.
-        test_dir = 'test'
-        test_path = os.path.join(DIR_TO_BACKUP, test_dir)
+        test_dir = 'subdir-to-copy'
+        test_path = os.path.join(FAKE_HOME_DIR, test_dir)
         os.mkdir(test_path)
-        copy_dir_if_valid(test_path, BACKUP_DIR)
+        copy_dir_if_valid(FAKE_HOME_DIR, BACKUP_DEST_DIR)
         assert os.path.isdir(test_path)
-        assert os.path.isdir(os.path.join(BACKUP_DIR, test_dir))
+        assert os.path.isfile(os.path.join(BACKUP_DEST_DIR, os.path.split(TEST_TEXT_FILE)[1]))
+        assert os.path.isdir(os.path.join(BACKUP_DEST_DIR, test_dir))
 
     @pytest.mark.parametrize('invalid', {".Trash", ".npm", ".cache", ".rvm"})
     def test_copy_dir_invalid(self, invalid):
         """
         Test that attempting to copy an invalid directory fails
         """
-        copy_dir_if_valid(invalid, DIR_TO_BACKUP)
-        assert not os.path.isdir(os.path.join(BACKUP_DIR, invalid))
+        copy_dir_if_valid(invalid, FAKE_HOME_DIR)
+        assert not os.path.isdir(os.path.join(BACKUP_DEST_DIR, invalid))
